+++ test/nn/kge/test_complex.py
@@ -6,13 +6,13 @@ from torch_geometric.nn import ComplEx
 def test_complex_scoring():
     model = ComplEx(5, 2, 1)
     model.node_emb.weight.data = torch.Tensor([[2.], [3.], [5.], [1.], [2.]])
-    model.node_emb_im.weight.data = torch.Tensor([[4.], [1.], [3.], [1.],
-                                                  [2.]])
+    model.node_emb_im.weight.data = torch.Tensor(
+        [[4.], [1.], [3.], [1.], [2.]])
     model.rel_emb.weight.data = torch.Tensor([[2.], [3.]])
     model.rel_emb_im.weight.data = torch.Tensor([[3.], [1.]])
 
-    score = model(torch.IntTensor([1, 3]), torch.IntTensor([1, 0]),
-                  torch.IntTensor([2, 4]))
+    score = model(torch.IntTensor([1, 3]), torch.IntTensor(
+        [1, 0]), torch.IntTensor([2, 4]))
 
     assert torch.all(torch.eq(score, torch.tensor([58, 8])))